Concrete AuctionId type.

+++ modules/dex/src/mock.rs
@@ -13,7 +13,7 @@ pub type AccountId = u128;
 pub type BlockNumber = u64;
 pub type Share = u128;
 pub type Amount = i128;
-pub type AuctionId = u64;
+pub type AuctionId = u32;
 
 pub const ALICE: AccountId = 1;
 pub const BOB: AccountId = 2;
